+++ .gitignore
@@ -1,1 +1,2 @@
 .vercel
+.DS_Store
