+++ client/package.json
@@ -2,7 +2,6 @@
   "name": "client",
   "version": "0.1.0",
   "private": true,
-  "proxy": "https://signezflask.up.railway.app",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
